Working out unity bugs

+++ src/plugin_array.cs
@@ -247,7 +247,7 @@ namespace Toy {
 					}
 
 					public object Call(Interpreter interpreter, Token token, List<object> arguments) {
-						return self.container.Count;
+						return (double)self.container.Count;
 					}
 
 					public override string ToString() { return "<Array function>"; }
